Generalize clipboard test fixture
It can now also be used by other tests in misc. Will be used for the
`version --copy` command.
